♻️ Refactor: my place pagination
