+++ src/components/MyRouteContent/MyRouteCardSection/MyRouteCardSectionPagination.tsx
@@ -1,27 +1,12 @@
 import { ChevronLeftIcon } from '@heroicons/react/16/solid';
 import { ChevronRightIcon } from '@heroicons/react/16/solid';
-import { useState } from 'react';
 
-const MyRouteCardSectionPagination = () => {
-  // const { offset, maxOffset, handleNextPage, handlePreviousPage } = useDashboardList();
-  const [offset, setOffset] = useState(1);
-  const maxOffset = 10;
+const MyRouteCardSectionPagination = ({ offset, setOffset, maxOffset, handlePreviousPage, handleNextPage }) => {
   const pageNumbers = [];
+
   for (let i = 1; i <= maxOffset; i++) {
     pageNumbers.push(i);
   }
-
-  const handleNextPage = () => {
-    if (offset < maxOffset) {
-      setOffset(offset + 1);
-    }
-  };
-
-  const handlePreviousPage = () => {
-    if (offset > 1) {
-      setOffset(offset - 1);
-    }
-  };
 
   return (
     <div className="mt-40 flex justify-center items-center">
@@ -33,7 +18,6 @@ const MyRouteCardSectionPagination = () => {
       >
         <ChevronLeftIcon />
       </div>
-      {/* <div className="flex gap-4 justify-center items-center"> */}
       {pageNumbers.map((number) => {
         if (
           number === 1 ||
@@ -66,7 +50,6 @@ const MyRouteCardSectionPagination = () => {
 
         return null;
       })}
-      {/* </div> */}
       <div
         className={`w-30 h-30 flex justify-center items-center cursor-pointer border-1 border-gray-30 rounded-4 ${
           offset === maxOffset && 'opacity-25 cursor-not-allowed'
