+++ src/components/MyRouteContent/MyRouteCardSection/index.tsx
@@ -14,7 +14,21 @@ const sortMenuList = ['데이트👩‍❤️‍👨', '가족👪', '혼자😏
 
 const MyRouteCardSection = ({ filteredData, setSearchValue }: filteredDataProps) => {
   const [sort, setSort] = useState<string>('인기순');
-  // const GRID_ROW = Math.ceil(filteredData.length / 4);
+  const [offset, setOffset] = useState(1);
+  const maxOffset = Math.ceil(filteredData.length / 9);
+
+  const handleNextPage = () => {
+    if (offset < maxOffset) {
+      setOffset(offset + 1);
+    }
+  };
+
+  const handlePreviousPage = () => {
+    if (offset > 1) {
+      setOffset(offset - 1);
+    }
+  };
+
   const handleClickSortMenu = (menus: string) => {
     setSort(menus);
     if (menus === '인기순') {
@@ -23,13 +37,15 @@ const MyRouteCardSection = ({ filteredData, setSearchValue }: filteredDataProps)
       menus.length <= 4 ? setSearchValue(menus.slice(0, -2)) : setSearchValue(menus.slice(0, -8));
     }
   };
-  // 리팩토링시 커스텀 훅으로 뺄것?
+
+  const paginatedData = filteredData.slice((offset - 1) * 9, offset * 9);
+
   return (
     <>
       <div className="flex mb-20 justify-between">
-        {sortMenuList.map((menus, idx) => (
+        {sortMenuList.map((menus, index) => (
           <input
-            key={idx}
+            key={index}
             className={`rounded-s h-36 min-w-110 font-bold bg-gray-10 py-5 px-15 text-15 cursor-pointer  ${
               sort === menus ? 'opacity-100' : 'opacity-50'
             }`}
@@ -39,17 +55,27 @@ const MyRouteCardSection = ({ filteredData, setSearchValue }: filteredDataProps)
           />
         ))}
       </div>
-      {filteredData.length !== 0 ? (
+      {paginatedData.length !== 0 ? (
         <>
-          <div className={`grid grid-cols-3 gap-12`}>
-            {filteredData.map((data, index) => (
+          <div className={`grid grid-cols-3 gap-12 h-700`}>
+            {paginatedData.map((data, index) => (
               <MyRouteListCard key={index} data={data} />
             ))}
           </div>
-          <MyRouteCardSectionPagination />
+          {offset > 1 && (
+            <MyRouteCardSectionPagination
+              offset={offset}
+              setOffset={setOffset}
+              maxOffset={maxOffset}
+              handleNextPage={handleNextPage}
+              handlePreviousPage={handlePreviousPage}
+            />
+          )}
         </>
       ) : (
-        <NoSearchData />
+        <div className="text-center">
+          <NoSearchData />
+        </div>
       )}
     </>
   );
